Time: 185 ms (22.56%), Space: 68.7 MB (60.88%) - LeetHub

+++ 1557-check-if-a-string-contains-all-binary-codes-of-size-k/1557-check-if-a-string-contains-all-binary-codes-of-size-k.java
@@ -10,6 +10,6 @@ class Solution {
             end++;
         }
         int length = set.size();
-        return length == Math.pow(2, k);
+        return length == (int)(Math.pow(2, k));
     }
 }
